Remove redundant args from dispatch event (#527)
