+++ typescript/utils/src/utils.ts
@@ -1,7 +1,8 @@
+import { arrayify, hexlify } from '@ethersproject/bytes';
 import { assert } from 'chai';
 import { ethers } from 'ethers';
 
-import { Address, Domain, HexString } from './types';
+import { Address, Domain, HexString, ParsedMessage } from './types';
 
 /*
  * Gets the byte length of a hex string
@@ -64,6 +65,22 @@ export const formatMessage = (
     [localDomain, senderAddr, destinationDomain, recipientAddr, body],
   );
 };
+
+/**
+ * Parse a serialized Abacus message from raw bytes.
+ *
+ * @param message
+ * @returns
+ */
+export function parseMessage(message: string): ParsedMessage {
+  const buf = Buffer.from(arrayify(message));
+  const origin = buf.readUInt32BE(0);
+  const sender = hexlify(buf.slice(4, 36));
+  const destination = buf.readUInt32BE(36);
+  const recipient = hexlify(buf.slice(40, 72));
+  const body = hexlify(buf.slice(72));
+  return { origin, sender, destination, recipient, body };
+}
 
 export function messageHash(message: HexString, leafIndex: number): string {
   return ethers.utils.solidityKeccak256(
